+++ src/nanonext.h
@@ -181,6 +181,7 @@ extern int nano_matchargs(SEXP);
 extern void nano_serialize(nano_buf *, SEXP);
 extern void nano_serialize_xdr(nano_buf *, SEXP);
 extern SEXP nano_unserialize(unsigned char *, const size_t);
+extern SEXP nano_hashToChar(unsigned char *, const size_t);
 extern SEXP rawToChar(unsigned char *, const size_t);
 extern void socket_finalizer(SEXP);